% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_halfeyeplot.R
\name{build_bot_halfeyeplot}
\alias{build_bot_halfeyeplot}
\title{Bot hap-pheno halfeye plot}
\usage{
build_bot_halfeyeplot(HapObject, hide_labels = T)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}

\item{hide_labels}{If TRUE, legend is hidden.}
}
\description{
build_bot_halfeyeplot() builds a vertical plot displaying the
phenotypic scores for each individual, grouped by haplotype, coloured by
metadata variable. Makes use of the $Indfile information from haplotype
object. It is an internal function called by crosshap_viz(), though can be
called separately to build a stand-alone plot.
}
