% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_jitterplot.R
\name{build_bot_jitterplot}
\alias{build_bot_jitterplot}
\title{Bot Hap-Pheno jitterplot}
\usage{
build_bot_jitterplot(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}
}
\description{
Internal function that creates a vertical jitterplot displaying the
phenotypic scores for each individual, grouped by haplotype. Makes use of the
$Indfile information. May be missing some axis to allow for 'crosshap'
stitching.
}
