% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_alleleplot.R
\name{build_left_alleleplot}
\alias{build_left_alleleplot}
\title{Left SNP-info alleleplot}
\usage{
build_left_alleleplot(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}
}
\description{
Internal function that plots the mean frequencies of all alleles for loci
within each marker group. Makes use of $Varfile information. The first
two snippets of code calculates the mean frequencies of each allele type for
each marker group, before plotting. May be missing some axis to allow for
'crosshap' stitching.
}
