% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_posplot.R
\name{build_left_posplot}
\alias{build_left_posplot}
\title{Left SNP posplot}
\usage{
build_left_posplot(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}
}
\description{
Internal function that plots the position of each SNP in each marker group
horizontally using small vertical lines. Makes use of the $MGfile position
information for each allele. The first snippet of code determines the labels
to use, as genomic position can be exceptionally high numbers that end up
being cut-off when default labeling is used. May be missing some axis to
allow for 'crosshap' stitching.
}
