% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_clusterplot.R
\name{build_right_clusterplot}
\alias{build_right_clusterplot}
\title{Right SNP cluster linkage plot}
\usage{
build_right_clusterplot(HapObject, hide_labels)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}

\item{hide_labels}{XX}
}
\description{
XXXInternal function that creates a horizontal phenoplot displaying the
difference in phenotype means between alternate and reference alleles for
each SNP loci, grouped by marker group. Makes use of the $Varfile phenotypic
information for each allele, first calculating the difference between alt/ref
before plotting. May be missing some axis to allow for 'crosshap' stitching.
}
