% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_jitterplot.R
\name{build_right_jitterplot}
\alias{build_right_jitterplot}
\title{Right SNP-Pheno jitterplot}
\usage{
build_right_jitterplot(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}
}
\description{
Internal function that creates a horizontal jitterplot displaying the
difference in phenotype means between alternate and reference alleles for
each SNP loci, grouped by marker group. Makes use of the $Varfile phenotypic
information for each allele, first calculating the difference between alt/ref
before plotting. May be missing some axis to allow for 'crosshap' stitching.
}
