% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{build_summary_tables}
\alias{build_summary_tables}
\title{XXSummary TablesXX}
\usage{
build_summary_tables(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}
}
\value{
A list containing two ggplot objects.
}
\description{
XXNeeds updating.XX
}
\examples{

if (FALSE) {
    XX
}

}
