% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_barplot.R
\name{build_top_metaplot}
\alias{build_top_metaplot}
\title{Top Metadata-hap Plot}
\usage{
build_top_metaplot(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}
}
\description{
Internal function that creates a vertical stacked barplot representing the
frequency of each haplotype combination, broken down by each categorical
metadata variable. Makes use of the $Hapfile pseudoSNP count information. May
be missing some axis to allow for 'crosshap' stitching.
}
