% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_viz.R
\name{clustree_viz}
\alias{clustree_viz}
\title{Clustering tree}
\usage{
clustree_viz(HapObject, type = "MG")
}
\arguments{
\item{HapObject}{A haplotyping object with a range of results from different
epsilons created by run_haplotyping()}

\item{type}{When type = "hap", nodes represent haplotype populations, when
type = "MG", nodes represent marker groups.}
}
\value{
A ggplot2 object.
}
\description{
clustree_viz() builds a clustering tree displaying changes in haplotype
assignment between individuals or changes in Marker Group assignment for
SNPs, across different epsilon values. This function is a `clustree` wrapper.
}
