% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.R
\name{create_pseudoSNP}
\alias{create_pseudoSNP}
\title{Create pseudoSNPs from SNP marker groups}
\usage{
create_pseudoSNP(MGfile, bin_vcf)
}
\arguments{
\item{MGfile}{SNP marker groups clustered using DBscan.}

\item{bin_vcf}{Input binary VCF for region of interest.}
}
\description{
Internal function that calls the most common allelic states for each SNP
marker group across individuals, returning dummy SNPs for each marker group
mimicking the binary vcf format. Options to be added for treatment of heterozygotes.
}
