% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hap_run_cluster}
\alias{hap_run_cluster}
\title{Run Clustering on Data}
\usage{
hap_run_cluster(vcf, LD, elon = c(0.5, 1, 1.5, 2))
}
\arguments{
\item{vcf}{Input vcf of region.}

\item{LD}{Pairwise R squared correlation matrix.}

\item{elon}{Epsilon values to search.}
}
\description{
Run Clustering on Data
}
