% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HapObject}
\alias{HapObject}
\title{Example Haplotype object}
\format{
A haplotype (S3) object containing results needed for haplotype visualization across five epsilon values (0.2,0.4,0.6,0.8,1)
\describe{
  \item{epsilon}{Epsilon value chosen for haplotyping with DBSCAN}
  \item{MGmin}{MGmin value (minPts) chosen for haplotyping with DBSCAN}
  \item{Hapfile}{Summary of Marker Groups defining haplotype combinations}
  \item{Indfile}{Haplotype assignments for individuals}
  \item{Varfile}{Marker Group assignments for SNPs, with additional calculated information}
  }
  @concept data
}
\usage{
HapObject
}
\description{
A haplotyping object created by run_haplotyping() for example cqProt-003 soy data
}
\keyword{datasets}
