% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LD}
\alias{LD}
\title{LD matrix}
\format{
A square matrix read in by read_LD()
}
\usage{
LD
}
\description{
A pairwise R^2 linkage matrix generated by PLINK for example cqProt-003 soy data
}
\concept{data}
\keyword{datasets}
