% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Domestication metadata}
\format{
A two-column tibble read in by read_metadata()
}
\source{
https://doi.org/10.1007/s00122-022-04045-8
}
\usage{
metadata
}
\description{
Metadata file with level of domestication for each individual in example cqProt-003 soy data
}
\concept{data}
\keyword{datasets}
