% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Example phenotype data}
\format{
A two-column tibble read in by read_pheno()
}
\source{
https://doi.org/10.1007/s00122-022-04045-8
}
\usage{
pheno
}
\description{
Seed protein scores for each individual in example cqProt-003 soy data
}
\concept{data}
\keyword{datasets}
