% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo2haps.R
\name{pseudo2haps}
\alias{pseudo2haps}
\title{Classify haplotype combinations and populations}
\usage{
pseudo2haps(pSNP, minHap)
}
\arguments{
\item{pSNP}{'pseudoSNPs' created by pseudo.R}

\item{minHap}{Minimum size (nIndividuals) to keep haplotype combinations}
}
\description{
Internal function that reads pseudoSNP data to identify and call the
frequency for characteristic marker group combinations. Defines common marker
group allele combinations as haplotypes and returns the haplotype of each
individual.
}
\examples{
pseudo2haps(het_pseudoSNP, 9)

}
