% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohaps.R
\name{pseudo_haps}
\alias{pseudo_haps}
\title{Identify haplotypes from SNP clusters}
\usage{
pseudo_haps(preMGfile, bin_vcf, minHap, LD, keep_outliers)
}
\arguments{
\item{preMGfile}{SNP clusters from DBscan.}

\item{bin_vcf}{Input binary VCF for region of interest.}

\item{minHap}{Minimum size (nIndividuals) to keep haplotype combinations}

\item{LD}{LD matrix input}

\item{keep_outliers}{whether to keep outlier SNPs in MGs}
}
\description{
Internal function that calls the most common allelic states for each SNP
marker group across individuals, returning dummy SNPs for each marker group
mimicking the binary vcf format. Options to be added for treatment of heterozygotes.
}
