% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_LD}
\alias{read_LD}
\title{Read correlation matrix between all SNPs in region}
\usage{
read_LD(LDin)
}
\arguments{
\item{LDin}{Correlation matrix output from PLINK}
}
\value{
A read in table of blah.
}
\description{
Read correlation matrix between all SNPs in region
}
