% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadat from two column text file without a header (Ind | Metadata)}
\usage{
read_metadata(Metain)
}
\arguments{
\item{Metain}{Input phenotype file}
}
\description{
Read metadat from two column text file without a header (Ind | Metadata)
}
