% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_pheno}
\alias{read_pheno}
\title{Read phenotype data from two column text file without a header (Ind | Pheno)}
\usage{
read_pheno(Phenoin)
}
\arguments{
\item{Phenoin}{Input phenotype file}
}
\description{
Read phenotype data from two column text file without a header (Ind | Pheno)
}
