% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF to tibble}
\usage{
read_vcf(VCFin)
}
\arguments{
\item{VCFin}{Input VCF}
}
\value{
A tibble.
}
\description{
Dashes,'-', in individual names are recoded to '.' for downstream
compatability.
}
\examples{
read_vcf(your_region.vcf)

}
