% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hdbscan_haplotyping.R
\name{run_hdbscan_haplotyping}
\alias{run_hdbscan_haplotyping}
\title{Cluster SNPs and identify haplotypes with HDBscan}
\usage{
run_hdbscan_haplotyping(
  vcf,
  LD,
  pheno,
  MGmin,
  minHap,
  hetmiss_as = "allele",
  metadata = NULL,
  keep_outliers = F
)
}
\arguments{
\item{vcf}{Input VCF for region of interest.}

\item{LD}{Pairwise correlation matrix of SNPs in region from PLINK.}

\item{pheno}{Input numeric phenotype data for each individual.}

\item{MGmin}{Minimum SNPs in marker groups, MinPts parameter for HDBscan.}

\item{minHap}{Minimum nIndividuals to keep haplotype combinations.}

\item{hetmiss_as}{Treat heterozygous missing './X' as missing or allele.}

\item{metadata}{Metadata input}

\item{keep_outliers}{Smoothing}
}
\description{
Performs density-based clustering of SNPs in region of interest to identify
marker groups. Defines haplotype combinations and classifies individuals
based on characteristic combinations of marker group alleles. Returns a
comprehensive haplotyping object, required to build clustering tree and
visualize haplotypes.
}
