% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagphenos.R
\name{tagphenos}
\alias{tagphenos}
\title{Calculate SNP phenotypic associations}
\usage{
tagphenos(MGfile, bin_vcf, pheno)
}
\arguments{
\item{MGfile}{SNP marker groups clustered using DBscan.}

\item{pheno}{Input numeric phenotype data for each individual.}

\item{vcf}{Input VCF for region of interest.}
}
\description{
Internal function that splits SNPs by their different allelic states and
calculates frequencies across the population. In addition, the mean phenotype
is calculated across all individuals that share a given allele for each loci.
}
