% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf}
\alias{vcf}
\title{VCF}
\format{
A VCF read in by read_vcf()
}
\source{
https://doi.org/10.1007/s00122-022-04045-8
}
\usage{
vcf
}
\description{
A VCF containing SNPs for example cqProt-003 soy data
}
\concept{data}
\keyword{datasets}
