% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_alleleplot.R
\name{build_left_alleleplot}
\alias{build_left_alleleplot}
\title{Left SNP-allele plot}
\usage{
build_left_alleleplot(HapObject, hide_labels = T)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}

\item{hide_labels}{If TRUE, legend is hidden.}
}
\value{
A ggplot2 object.
}
\description{
build_left_alleleplot() builds a horizontal plot displaying mean allelic
frequencies (reference/alternate/missing/heterozygous) of all SNP loci,
grouped by marker group. Makes use of $Varfile information from a HapObject
created by run_haplotyping(). This is an internal function called by
crosshap_viz(), though can be called separately to build a stand-alone plot.
}
\examples{

if (FALSE) {
     build_left_alleleplot(Haplotypes_MGmin30_E0.6, hide_labels = F)
}

}
