% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_posplot.R
\name{build_left_posplot}
\alias{build_left_posplot}
\title{Left SNP position plot}
\usage{
build_left_posplot(HapObject, hide_labels = T)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}

\item{hide_labels}{If TRUE, legend is hidden.}
}
\description{
build_left_alleleplot() builds a horizontal plot displaying the chromosomal
position of each SNP locus, grouped by marker group.Makes use of the $MGfile
position from haplotype object. It is an internal function called by
crosshap_viz(), though can be called separately to build a stand-alone plot.
}
