% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_dotplot.R
\name{build_mid_dotplot}
\alias{build_mid_dotplot}
\title{Middle MG/hap dot plot}
\usage{
build_mid_dotplot(HapObject, hide_labels)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}

\item{hide_labels}{}
}
\value{
A ggplot2 object.
}
\description{
build_mid_dotplot() builds a central dot plot displaying the relationship
between haplotype combinations and the characteristic marker group alleles
that define them. Makes use of the $Hapfile information from a haplotype
object. This is an internal function called by crosshap_viz(), though can
be called separately to build a stand-alone plot (can be useful when patched
to a a peripheral plot).
}
\examples{

if (FALSE) {
     build_mid_dotplot(Haplotypes_MGmin30_E0.6, hide_labels = F)
}

}
