% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_dotplot.R
\name{build_mid_dotplot}
\alias{build_mid_dotplot}
\title{Middle MG/Hap dotplot}
\usage{
build_mid_dotplot(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by crosshap::run_haplotyping}
}
\description{
Internal function that creates central dotplot displaying the relationship
between haplotype combinations and the characteristic marker groups they
possess. Makes use of the $Hapfile 'pseudoSNP' information. Required as an
anchor for all peripheral plots in the 'crosshap' visualization.
}
