% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_clusterplot.R
\name{build_right_clusterplot}
\alias{build_right_clusterplot}
\title{Right SNP cluster linkage plot}
\usage{
build_right_clusterplot(HapObject, hide_labels = F)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}

\item{hide_labels}{If TRUE, legend is hidden.}
}
\description{
build_right_jitterplot() builds a horizontal plot displaying the mean
pairwise R^2 linkage between each SNP and all other SNPs in its marker group,
grouped by marker group, coloured by alternate allele frequency. Makes use of
the $Varfile and $MGfile information from haplotyping object. It is an
internal function called by crosshap_viz(), though can be called separately
to build a stand-alone plot.
}
