% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{build_summary_tables}
\alias{build_summary_tables}
\title{Hap/MG summary tables}
\usage{
build_summary_tables(HapObject, epsilon)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}

\item{epsilon}{Epsilon to visualize haplotyping results for.}
}
\value{
A list containing two TableGrob objects.
}
\description{
build_summary_tables() builds summary tables for each haplotype and Marker
Group with some of the information shown in the peripheral crosshap plots.
It is an internal function called by crosshap_viz(), though can be called
separately to build stand-along grob tables.
}
