% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{build_summary_tables}
\alias{build_summary_tables}
\title{Hap/MG summary tables}
\usage{
build_summary_tables(HapObject)
}
\arguments{
\item{HapObject}{Haplotype object created by run_haplotyping().}
}
\value{
A list containing two TableGrob objects.
}
\description{
build_summary_tables() builds summary tables for each haplotype and Marker
Group with some of the information shown in the peripheral crosshap plots.
It is an internal function called by crosshap_viz(), though can be called
separately to build stand-along grob tables.
}
\examples{

if (FALSE) {
     build_summary_tables(Haplotypes_MGmin30_E0.6)
}

}
