% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosshap-package.R
\docType{package}
\name{crosshap-package}
\alias{crosshap}
\alias{crosshap-package}
\title{crosshap: Local Haplotype Clustering and Visualization}
\description{
A local haplotyping visualization toolbox to capture major patterns of co-inheritance between clusters of linked variants, whilst connecting findings to phenotypic and demographic traits across individuals. 'crosshap' enables users to explore and understand genomic variation across a trait-associated region. For an example of successful local haplotype analysis, see Marsh et al. (2022) \doi{10.1007/s00122-022-04045-8}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jacobimarsh.github.io/crosshap/}
  \item Report bugs at \url{https://github.com/jacobimarsh/crosshap/issues}
}

}
\author{
\strong{Maintainer}: Jacob Marsh \email{jake.marsh@live.com.au} (\href{https://orcid.org/0000-0003-3734-2023}{ORCID})

Authors:
\itemize{
  \item Brady Johnston \email{brady.johnston@me.com} (\href{https://orcid.org/0000-0001-6301-2269}{ORCID})
  \item Jakob Petereit \email{petinho86@gmail.com} (\href{https://orcid.org/0000-0003-2159-0380}{ORCID})
}

}
\keyword{internal}
