% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosshap-package.R
\docType{package}
\name{crosshap-package}
\alias{crosshap}
\alias{crosshap-package}
\title{crosshap: Haplotype clustering and visualization}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jacobimarsh.github.io/crosshap/}
}

}
\author{
\strong{Maintainer}: Jacob Marsh \email{jacob.marsh@research.uwa.edu.au}

}
\keyword{internal}
