% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_LD}
\alias{read_LD}
\title{Read LD correlation matrix to tibble}
\usage{
read_LD(LDin)
}
\arguments{
\item{LDin}{Correlation matrix output from PLINK}
}
\value{
A tibble.
}
\description{
Read LD correlation matrix to tibble
}
\examples{
$ plink --r2 square --vcf your_region.vcf
read_LD(plink.out)

}
