% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata to tibble}
\usage{
read_metadata(Metain)
}
\arguments{
\item{Metain}{Input phenotype file}
}
\value{
A tibble.
}
\description{
Requires two column text file without a header (Ind | Metadata)
}
\examples{

if (FALSE) {
     read_metadata(country_of_origin.txt)
}

}
