% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_pheno}
\alias{read_pheno}
\title{Read phenotype data to tibble}
\usage{
read_pheno(Phenoin)
}
\arguments{
\item{Phenoin}{Input phenotype file}
}
\value{
A tibble.
}
\description{
Requires two column text file without a header (Ind | Pheno)
}
