% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_viz.R
\name{run_clustree}
\alias{run_clustree}
\title{Visualize clusters}
\usage{
run_clustree(epsilon = c(0.4, 0.8, 1.2, 1.6, 2), MGmin, pheno, type = "MG")
}
\arguments{
\item{epsilon}{Epsilon values to search.}

\item{pheno}{Input numeric phenotype data for each individual.}

\item{type}{Hap or MG clustree}
}
\description{
This function reads haplotype objects calculated for a range of epsilon
values. It calculates the mean phenotypic value for each haplotype group
and rearranges the data into a single sheet. Visualization is conducted using
the clustree package, with labeling manually added afterward.
}
