% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_viz.R
\name{run_clustree}
\alias{run_clustree}
\title{Haplotype clustering tree}
\usage{
run_clustree(
  pheno,
  epsilon = c(0.2, 0.4, 0.6, 0.8, 1),
  MGmin = 30,
  type = "MG"
)
}
\arguments{
\item{pheno}{Input numeric phenotype data for each individual.}

\item{epsilon}{Epsilon values passed through run_haplotyping().}

\item{MGmin}{MGmin values passed through run_haplotyping().}

\item{type}{When type = "hap", nodes represent haplotype populations, when
type = "MG", nodes represent marker groups.}
}
\value{
A ggplot2 object.
}
\description{
run_clustree() builds a clustering tree displaying changes in haplotype
assignment between individuals or changes in marker group assignment for
SNPs, across different epsilon values provided to run_haplotyping(). Makes
use of $Indfile and $Varfile information from HapObjects created by
run_haplotyping(). This function is a clustree wrapper.
}
\examples{

if (FALSE) {
     run_clustree(MGmin = 30, pheno)
}

}
