% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_haplotyping_wrapper.R
\name{run_haplotyping}
\alias{run_haplotyping}
\title{Cluster SNPs and identify haplotypes}
\usage{
run_haplotyping(
  vcf,
  LD,
  pheno,
  epsilon = c(0.5, 1, 1.5, 2, 2.5, 3),
  MGmin,
  minHap
)
}
\arguments{
\item{vcf}{Input VCF for region of interest.}

\item{LD}{Pairwise correlation matrix of SNPs in region from PLINK.}

\item{pheno}{Input numeric phenotype data for each individual.}

\item{epsilon}{Epsilon values for clustering SNPs with DBscan.}

\item{MGmin}{Minimum SNPs in marker groups, MinPts parameter for DBscan.}

\item{minHap}{Minimum nIndividuals to keep haplotype combinations}
}
\description{
Performs density-based clustering of SNPs in region of interest to identify
marker groups. Defines haplotype combinations and classifies individuals
based on characteristic combinations of marker group alleles. Returns a
comprehensive haplotyping object, required to build clustering tree and
visualize haplotypes.
}
