% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagphenos.R
\name{tagphenos}
\alias{tagphenos}
\title{Calculate SNP phenotypic associations}
\usage{
tagphenos(MGfile, bin_vcf, pheno, het_phenos = FALSE)
}
\arguments{
\item{MGfile}{SNP marker groups clustered using DBscan.}

\item{bin_vcf}{Binary VCF for region of interest reformatted by
run_haplotyping().}

\item{pheno}{Input numeric phenotype data for each individual.}

\item{het_phenos}{When FALSE, phenotype associations for SNPs are calculated
from reference and alternate allele individuals only, when TRUE, heterozygous
individuals are included assuming additive allele effects.}
}
\value{
Returns intermediate of haplotype object.
}
\description{
tagphenos() reports the frequency of allele types for each SNP and calculates
phenotype associations for the different alleles, before returning this
information in a $Varfile in a HapObject. This is an internal function that
is not intended for external use.
}
